requirements.txt
asgiref==3.2.10
config==0.5.0.post0
confusable-homoglyphs==3.2.0
coverage==5.2.1
dj-database-url==0.5.0
Django==3.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-registration==3.1
djangorestframework==3.11.1
gunicorn==20.0.4
install==1.3.3
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
